requirements.txt
# user pip install -r requirements.txt -i https://pypi.douban.com/simple
############################################################################################

pysimplegui
pyinstaller
baidu-aip
chardet
pillow
